Moved valid date range function into it's own file to be accessibly within the dynamic pages.
