+++ pages/calendar/[...date].tsx
@@ -1,7 +1,8 @@
-import React, { useEffect, useState } from "react";
+import React, { useEffect, useState, useContext } from "react";
 import { Box } from "@chakra-ui/react";
 import { useRouter } from "next/router";
 import { endOfMonth, getDay } from "date-fns";
+import findValidDateRange from "../../lib/findValidDateRange"
 import ErrorPage from "next/error";
 import Calender from "../../components/calender";
 import { CalenderContextProvider } from "../../contexts/CalenderContext";
@@ -10,6 +11,9 @@ import { StickersContextProvider } from "../../contexts/StickerContext";
 const DateRoute: React.FC<unknown> = () => {
   const router = useRouter();
   const { date: slug } = router.query;
+
+  const validDateRange = findValidDateRange();
+  const { start: validStart, end: validEnd } = validDateRange;
 
   const [date, setDate] = useState<UpdateCalendarProps | null>(null);
 
@@ -54,7 +58,11 @@ const DateRoute: React.FC<unknown> = () => {
     return date;
   };
 
-  useEffect(() => {
+  // const validateDateRange = () => {
+
+  // }
+
+  useEffect(() => {2
     if (slug && slug.length === 1 && slug[0] !== "now") {
       setError(true);
       return console.warn("improper date input");
@@ -75,7 +83,9 @@ const DateRoute: React.FC<unknown> = () => {
         });
       }
     }
-  }, [slug]);
+
+    console.info("Context:", calenderContext)
+  }, [slug, calenderContext]);
 
   if (router.isFallback) {
     return <ErrorPage statusCode={404} />;
